requirements.txt
#Doesn't include pytorch
torchvision
Pillow
argparse
numpy
nltk
gensim
matplotlib
